requirements.txt
crawl4ai>=0.7.4
pytrends
openai
python-dotenv
httpx
